slant-backend/requirements.txt
flask==3.1.0
flask-cors==5.0.0
pandas==2.1.4
markdown==3.5.2
langchain==0.3.20
doppler-sdk==1.3.0
psycopg2==2.9.10
langchain_anthropic==0.3.9
langchain_openai==0.3.7
langchain_core==0.3.41
pinecone==5.4.2
requests==2.32.0
anthropic==0.49.0
numpy==1.26.3
sqlalchemy==2.0.36
selenium==4.19.0
beautifulsoup4==4.12.2
tqdm==4.65.0
langchain-community==0.3.19
gunicorn==23.0.0
langgraph==0.3.5
asyncio==3.4.3
slack-sdk==3.33.5
flipside==2.1.0
langchain_postgres==0.0.13
psycopg[binary]==3.2.6
tabulate==0.9.0
tavily-python==0.5.4